add switch logic
